rating and comment added

+++ Fitness/Models/UserRate.cs
@@ -11,7 +11,7 @@ public partial class UserRate
 
     public int? ChallengeId { get; set; }
 
-    public short? Rate { get; set; }
+    public int? Rate { get; set; }
 
     public virtual Challenge? Challenge { get; set; }
 }
